+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultButtonPanelBuilder.java
@@ -130,7 +130,7 @@ final class DefaultButtonPanelBuilder extends AbstractControlPanelBuilder<JPanel
 
 		@Override
 		void onToggleControl(ToggleControl toggleControl) {
-			JToggleButton button = toggleButtonBuilder.toggleControl(toggleControl).build();
+			JToggleButton button = toggleButtonBuilder.toggle(toggleControl).build();
 			if (buttonGroup != null) {
 				buttonGroup.add(button);
 			}