+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultToolBarBuilder.java
@@ -102,7 +102,7 @@ final class DefaultToolBarBuilder extends AbstractControlPanelBuilder<JToolBar, 
 		@Override
 		void onToggleControl(ToggleControl toggleControl) {
 			toolBar.add(toggleButtonBuilder
-							.toggleControl(toggleControl)
+							.toggle(toggleControl)
 							.build());
 		}