+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/ToggleButtonBuilder.java
@@ -34,13 +34,13 @@ public interface ToggleButtonBuilder<C extends JToggleButton, B extends ToggleBu
 	 * @param toggleControl the toggle control to base this toggle button on
 	 * @return this builder instance
 	 */
-	B toggleControl(ToggleControl toggleControl);
+	B toggle(ToggleControl toggleControl);
 
 	/**
 	 * @param toggleControlBuilder the builder for the toggle control to base this toggle button on
 	 * @return this builder instance
 	 */
-	B toggleControl(Control.Builder<ToggleControl, ?> toggleControlBuilder);
+	B toggle(Control.Builder<ToggleControl, ?> toggleControlBuilder);
 
 	/**
 	 * @param <C> the component type